improves glass, visuals and method of saying things

+++ code/modules/research/fabricators.dm
@@ -355,8 +355,11 @@
 		being_built.forceMove(get_turf(output))
 		being_built.anchored = 0
 		if(arcanetampered)
-			qdel(being_built)
-			being_built = new /obj/item/weapon/bikehorn/rubberducky(get_turf(output)) // BONUS DUCKS! No material refunds
+			if(prob(90))
+				qdel(being_built)
+				being_built = new /obj/item/weapon/bikehorn/rubberducky(get_turf(output)) // BONUS DUCKS! No material refunds
+			else
+				being_built.arcane_act(usr)
 		visible_message("[bicon(src)] \The [src] beeps: \"Successfully completed \the [being_built.name].\"")
 		being_built = null
 		last_made = part
